Move error generation method to abstract controller
