+++ library/Zend/Mvc/Controller/AbstractActionController.php
@@ -12,7 +12,6 @@ namespace Zend\Mvc\Controller;
 use Zend\Http\Response as HttpResponse;
 use Zend\Mvc\Exception;
 use Zend\Mvc\MvcEvent;
-use Zend\View\Model\ConsoleModel;
 use Zend\View\Model\ViewModel;
 
 /**
@@ -85,33 +84,5 @@ abstract class AbstractActionController extends AbstractController
         $e->setResult($actionResponse);
 
         return $actionResponse;
-    }
-
-    /**
-     * Create an HTTP view model representing a "not found" page
-     *
-     * @param  HttpResponse $response
-     * @return ViewModel
-     */
-    protected function createHttpNotFoundModel(HttpResponse $response)
-    {
-        $response->setStatusCode(404);
-        return new ViewModel(array(
-            'content' => 'Page not found',
-        ));
-    }
-
-    /**
-     * Create a console view model representing a "not found" action
-     *
-     * @param  \Zend\Stdlib\ResponseInterface $response
-     * @return ConsoleModel
-     */
-    protected function createConsoleNotFoundModel($response)
-    {
-        $viewModel = new ConsoleModel();
-        $viewModel->setErrorLevel(1);
-        $viewModel->setResult('Page not found');
-        return $viewModel;
     }
 }
